Add Features and rename scripts
Add the search feature which shows results of a given query by the user.
This results may be filtered by post type, or show users or tags. For this
reason, the tags page as well as its related functions has been dropped.
Add a profile page that shows the user's posts as any other profile
page.
Add settings page that allows users to change their information.
Add navbar that provides access to the aforementioned features.
Change filenames of scripts to be the same as the page that uses it, or
its main function.
Remove `request` dependency from server as it has entered into
maintenance mode.
Add `node-fetch` dependency to server to replace `request`. This has the
added benefit of being lighter.
Fix smaller bugs.
Add smaller features.